+++ packages/tswebextension/src/lib/common/constants.ts
@@ -1,1 +1,2 @@
 export const USER_FILTER_ID = 0;
+export const BACKGROUND_TAB_ID = -1;