+++ packages/tswebextension/src/lib/common/index.ts
@@ -8,3 +8,4 @@ export * from './stealth-helper';
 export * from './content-script/send-app-message';
 export * from './request-type';
 export * from './error';
+export * from './constants';